+++ scripts/tests/calibnet_export_check.sh
@@ -20,8 +20,6 @@ $FOREST_CLI_PATH snapshot export --tipset 0 --format "$format"
 echo "Exporting zstd compressed snapshot in $format format"
 $FOREST_CLI_PATH snapshot export --format "$format"
 
-$FOREST_CLI_PATH shutdown --force
-
 for f in *.car.zst; do
   echo "Inspecting archive info $f"
   $FOREST_TOOL_PATH archive info "$f"
@@ -39,3 +37,36 @@ for f in *.car.zst; do
   echo "Validating CAR file $f"
   $FOREST_TOOL_PATH snapshot validate "$f"
 done
+
+echo "Cleaning up the last snapshot"
+rm --force --verbose ./*.{car,car.zst,sha256sum}
+
+output=$($FOREST_CLI_PATH snapshot export-status --format json)
+is_exporting=$(echo "$output" | jq -r '.exporting')
+echo "Testing that no export is in progress"
+if [ "$is_exporting" == "true" ]; then
+  exit 1
+fi
+
+echo "Exporting zstd compressed snapshot"
+$FOREST_CLI_PATH snapshot export
+sleep 1
+
+output=$($FOREST_CLI_PATH snapshot export-status --format json)
+is_exporting=$(echo "$output" | jq -r '.exporting')
+echo "Testing that export is in progress"
+if [ "$is_exporting" == "false" ]; then
+  exit 1
+fi
+
+$FOREST_CLI_PATH snapshot export-cancel
+
+output=$($FOREST_CLI_PATH snapshot export-status --format json)
+is_exporting=$(echo "$output" | jq -r '.exporting')
+is_cancelled=$(echo "$output" | jq -r '.cancelled')
+echo "Testing that export has been cancelled"
+if [ "$is_exporting" == "true" ] || [ "$is_cancelled" == "false" ]; then
+  exit 1
+fi
+
+$FOREST_CLI_PATH shutdown --force